# backend-playground/data-or-search/redis-playground/docs/swagger.yaml
basePath: /v1
definitions:
  entity.User:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      username:
        example: johndoe
        type: string
    type: object
  github_com_agusheryanto182_redis-playground_internal_controller_restapi_v1_response.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  v1.Login:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: secret123
        type: string
    required:
    - email
    - password
    type: object
  v1.Register:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: secret123
        minLength: 6
        type: string
      username:
        example: johndoe
        maxLength: 255
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  v1.Token:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Multi-domain clean architecture template with translation, user, and
    task management
  title: Go Clean Template API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and get JWT token
      operationId: login
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Token'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_agusheryanto182_redis-playground_internal_controller_restapi_v1_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_agusheryanto182_redis-playground_internal_controller_restapi_v1_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_agusheryanto182_redis-playground_internal_controller_restapi_v1_response.Error'
      summary: Login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      operationId: register
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.Register'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_agusheryanto182_redis-playground_internal_controller_restapi_v1_response.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_agusheryanto182_redis-playground_internal_controller_restapi_v1_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_agusheryanto182_redis-playground_internal_controller_restapi_v1_response.Error'
      summary: Register
      tags:
      - auth
  /user/profile:
    get:
      description: Get current user profile
      operationId: profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_agusheryanto182_redis-playground_internal_controller_restapi_v1_response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_agusheryanto182_redis-playground_internal_controller_restapi_v1_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_agusheryanto182_redis-playground_internal_controller_restapi_v1_response.Error'
      security:
      - BearerAuth: []
      summary: Get profile
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"